Move Signer enum from indexer-types to runtime/wit/resources
