+++ core/indexer-types/src/lib.rs
@@ -225,56 +225,6 @@ pub struct Block {
     pub transactions: Vec<Transaction>,
 }
 
-#[derive(Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize, TS)]
-#[ts(export, export_to = "../../../kontor-ts/src/bindings.d.ts")]
-pub enum Signer {
-    Core(Box<Signer>),
-    XOnlyPubKey(String),
-    ContractId {
-        #[ts(type = "number")]
-        id: i64,
-        id_str: String,
-    },
-    Nobody,
-}
-
-impl Signer {
-    pub fn new_contract_id(id: i64) -> Self {
-        Self::ContractId {
-            id,
-            id_str: format!("__cid__{}", id),
-        }
-    }
-
-    pub fn is_core(&self) -> bool {
-        matches!(self, Signer::Core(_))
-    }
-}
-
-impl core::ops::Deref for Signer {
-    type Target = str;
-
-    fn deref(&self) -> &str {
-        match self {
-            Self::Nobody => "nobody",
-            Self::Core(_) => "core",
-            Self::XOnlyPubKey(s) => s,
-            Self::ContractId { id_str, .. } => id_str,
-        }
-    }
-}
-
-impl core::fmt::Display for Signer {
-    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
-        match self {
-            Self::Nobody => write!(f, "nobody"),
-            Self::Core(_) => write!(f, "core"),
-            Self::XOnlyPubKey(s) => write!(f, "{s}"),
-            Self::ContractId { id_str, .. } => write!(f, "{id_str}"),
-        }
-    }
-}
-
 #[derive(Debug, Clone, Hash, PartialEq, Eq, Serialize, Deserialize, TS)]
 #[ts(export, export_to = "../../../kontor-ts/src/bindings.d.ts")]
 pub struct UtxoRef {
@@ -330,15 +280,6 @@ impl core::str::FromStr for HolderRef {
     }
 }
 
-impl From<&Signer> for HolderRef {
-    fn from(signer: &Signer) -> Self {
-        match signer {
-            Signer::XOnlyPubKey(s) => HolderRef::XOnlyPubkey(s.clone()),
-            Signer::ContractId { id_str, .. } => HolderRef::ContractId(id_str.clone()),
-            Signer::Core(_) => HolderRef::Core,
-            Signer::Nobody => HolderRef::Core,
-        }
-    }
 }
 
 #[derive(Debug, Clone)]
